fiji_k_means/requirements.txt
opencv-python
numpy
scikit-learn
Pillow
roifile

